+++ examples/love2d Threading Example/GuiManagerDebug/init.lua
@@ -1,5 +1,4 @@
 utf8 = require("utf8")
-_defaultfont = love.graphics.getFont()
 gui = {}
 gui.__index = gui
 gui.TB={}
@@ -7,14 +6,15 @@ gui.Version="VERSION" -- Is it really ready for release?
 _GuiPro={GBoost=true,hasDrag=false,DragItem={},Children={},Visible=true,count=0,x=0,y=0,height=0,width=0,update=function(self) local things=GetAllChildren2(self) UpdateThings(things) end,draw=function(self) local things=GetAllChildren(self) DrawThings(things) end,getChildren=function(self) return self.Children end}
 _GuiPro.Clips={}
 _GuiPro.rotate=0
+_defaultfont = love.graphics.setNewFont(12)
 setmetatable(_GuiPro, gui)
 function gui:LoadInterface(file)
 	local add=".int"
 	if string.find(file,".",1,true) then add="" end
-	if love.filesystem.exists(file..add) then
+	if love.filesystem.getInfo(file..add) then
     a,b=pcall(love.filesystem.load(file..add))
 		if a then
-			print("Loaded: "..file)
+			--print("Loaded: "..file)
 		else
 			print("Error loading file: "..file)
       print(a,b)
@@ -34,7 +34,6 @@ function gui.LoadAll(dir)
 end
 -- Start Of Load
 
---gui.LoadAll("GuiManager/LibCore")
 gui.LoadAll("GuiManager/Core")
 gui.LoadAll("GuiManager/Image-Animation")
 gui.LoadAll("GuiManager/Frame")
@@ -42,10 +41,8 @@ gui.LoadAll("GuiManager/Item")
 gui.LoadAll("GuiManager/Misc")
 gui.LoadAll("GuiManager/Text")
 gui.LoadAll("GuiManager/Drawing")
-gui.LoadAll("GuiManager/Combos")
---gui.LoadAll("GuiManager/WIP")
-multi.boost=2
 
+multi.boost=2
 -- End of Load
 gui:respectHierarchy()
 _GuiPro.width,_GuiPro.height=love.graphics.getDimensions()